requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
Pillow==8.0.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
